Adding dataIdentification legalConstraints
legalConstraints are complex elements in that they are free-form in nature. We constrain the options available to an access constraint with an Inspire Limitation on Public Access term, and a use constraint with a copyright notice (if UK OGL v3) and/or a required citation.

Each of these supported use cases are optional.