Tidying up
Adding consistency with element classes.

All elements should appended to a parent element, rather than the root element, to ensure classes can be reused in different places (e.g. language).

When passing attributes to another class, the specific properties should be passed directly, rather than the receving element needing to select their properties. The only exception should be when an element class only uses a single attribute, and so passing that attribute would no longer be a dict.

Replacing ‘isinstance’ with ‘type’ to ensure ‘datetimes’ are not considered ‘dates’.