Implementing minimal config class
Currently accepts a complete dictionary of configuration and storing this as an attributes property.

This attributes dict is then used directly in the metadata class as if it was passed directly as a dict (as we did before). This is transitional and intended to be removed in the longer term.

We can then change how this attributes dict is generated, without needing to change how the attributes are used in the metadata element classes.