Refactoring ISO 19115 methods into a separate module
As a first step to adding other standards (such as a minimal schema), this moves all the ISO 19115 specific element classes and configuration into a separate module.

This leaves a much smaller set of generic classes for configs, records, namespaces and elements.

These classes are extended in the ISO 19115 module to add relevant concepts and elements.

This work extends to the JSON Schema, which now has a minimal base schema and ISO 19115 schema. Currently both schemas are standalone, but will be unified in the future by extending the base schema.

Tests have also been refactored to match this new structure.

The Flask application has been refactored into it’s own module, outside of the main project, since this project is to be used as a library, rather than a standalone service.

As part of general refactoring, the metadataRecord class has been extended to generate an XML document as a string for ease of use.