Adding minimal ISO 19115 test record
Following the same pattern developed for the test standard.

The minimal record is based on, but not a copy of, a minimal record from the validator project. Therefore this minimal record may not be valid, however this doesn’t particularly matter yet (since these are just tests) and can easily be changed when we move to validation.