Adding Flask command to output internal config schemas
So they can be uploaded to the metadata standards static site.

This changes the way that record configurations are validated, in that you now need to proactively do this. In reality it will still be automatically checked as part of instantiating a metadata record object, rather than a metadata record config object.