Fixing captured test record
These static records are used to ensure whole records are produced correctly. If the config used for a test changes then the record generated and compared against this static file will be different.

This change wasn’t picked up because of the order tests are ran (I got lazy and didn’t wait for them to all to finish).