Fixing support for distributors in ISO 19115
Previously we only created a distribution info element if there were either distribution formats or distribution transfer options.

This meant in cases where there are only distributors, they are omitted because their parent distribution info element was not made.

In reality this is unlikely to be the case, but is supported by the minimal implementations of the ISO 19115, Inspire and UK PDC Discovery standard/profiles so should work here too.