Refactoring JSON Schemas into files
Previously these were embedded in metadata standard classes because it wasn’t easy to reference them as files when installed as a package in other applications.

The `importlib` package resolves this problem and lets us store the schemas as standalone files. This also makes it easier to copy them to S3 without needing a custom CLI task.

The instructions for creating a new standard have been updated accordingly.

Tests have been added to ensure the configuration schemas are loaded and used correctly.