Adding support for generating configuration objects from records
I.e. the reverse of generating a record from a configuration object.

This support is intended to be lossless, meaning you could generate a record from a config, generate a config from that record and both config objects would be the same. To ensure/test this, some test configurations have been extected slightly.

The base MetadataRecord class has been extended to allow either a configuration object, or a  record string (typically loaded from a file) to be used.

Each element class has been extended to add a `make_config()` method. This is similar to `make_element()` but use XPaths to extract information to create a configuration object.

Tests and README examples have been updated as needed to test and demonstrate this new support.

As a first implementation, there are likely to be some rough edges, both in support for edge-cases and in possibly inefficent/inelegant code. Notably, coverage exemptions have been used in the ISO 19115 classes for defensive checks (i.e. if element is not populated don’t add to config object). These aren’t ideal but aren’t expected to ever be used. Adding tests to cover this code properly is possible but is not a priority for now. It’s expected some of these edges and/or testing limitations will be smoothed down and/or addressed in time.