Extracting common ISO elements to a standalone module
In preparation for adding ISO 19115-2, moving most of the classes defined in the current ISO 19115-1 (which is not actually 19115-1, but effectively 19115-0. When we add support for the real ‘-1’ we will refer to it as ‘-3’) to a separate ‘common’ namespace.

This includes ensuring all classes use allow the root xpath element (e.g. ‘/gmd:MD_Metadata’) to be configurable, as this changes between ISO variants.

This will prevent the new ISO 19115-2 classes depending on anything from ISO 19115-1 but rather both will depend on a common set of classes that implement the content of both standards.