Updating Black dependency
This was necessary due to a mismatch between the version of Black I was running locally and the version installed in the container.

As I couldn’t downgrade my local version I updated the version in the container. I’ve pinned Black on my local machine now so this shouldn’t happen again unless we choose to update both together.