Preparing for ISO 19115 V3 schemas
New schema and record config class added for both ISO 19115-1 and I9115-2 (we refer to ISO 19115:2003 as 19115-1, and will refer to I9115-1:2014 as 19115-3).

Initially this v3 schema is identical to v2.

The metadata record classes, tests, documentation and other resources have been updated to start using this v3 schema, meaning the ‘main’ branch of this repo is using an unreleased schema as it is developed and tested.