Changing lineage to an object in preparation for adding sources
The lineage configuration property is currently a string, which maps to a lineage statement in the ISO XML.

In the future (#73) we want to support lineage sources, in addition to statements. This will require chnaging the configuration property from a string to an object, which is a breaking change.

As other breaking configuration changes are being made in the V3 configuration schema, which is due to be released soon, this commit performs the breaking change necessary for the lineage configuration property (string -> object).

This commit does not add support for sources, which will come later, and refactoring will be needed within the code to create proper Source and Statement classes.

This commit is being made in the interests of expediency, in order to prevent a bottle-neck with other projects/initiatives.